clean up Alonzo tests Plutus & Pool and Smash
